mini project
This project is focused on customer account services in bank, so it is named “Customer Account Bank Management System”. Here, you can create a new account, update information of an existing account, view and manage transactions, check the details of an existing account, remove existing account and view customers’ list.
Overall, with this project, you can perform banking activities like in a REAL bank. Bank management mini project in C is a console application without graphics. It is compiled in Code::Blocks with gcc compiler.