periodic table
This project will help you to understand file handling in C i.e. creating a file and accessing the stored data in the file, modifying and removing the stored data. It will also help you to understand the use of functions as well as different parameters of C programming language.

The main functions used in designing the Modern Periodic Table project are:

void add(): This function is used to input or add the information of new element to the program.
void explor(): This function is used to explore the stored information in the file created.
void mainscreen(): It is included in source code of project file in order to print the text style and to control its color.
void mainscreen(): This function is used to print the main screen or menu of the project.
The key features of Modern Periodic Table mini project in C are briefly described below:

Storage of Element Information: In the project, you can add any new element with its name, symbol, atomic number, atomic weight and its some important properties. When new element information is to be added to this Modern Periodic Table, you have to enter 1 in the main menu and input information in given format. These information are stored in file created on the hard disk of computer by program itself.
Exploration of element Information: Another main function of this project is to explore or to display the stored information. You can search an element by using any of the following method:
By name of element
By symbol of element
By atomic number of element
By atomic weight of element